Only expose needed properties in IPool
