+++ src/index.ts
@@ -1,15 +1,15 @@
 export { AbstractPool } from './pools/abstract-pool'
-export type { PoolOptions } from './pools/abstract-pool'
-export { DynamicClusterPool } from './pools/cluster/dynamic'
-export { FixedClusterPool } from './pools/cluster/fixed'
-export type { ClusterPoolOptions } from './pools/cluster/fixed'
 export type {
   ErrorHandler,
   ExitHandler,
-  IPool,
   IWorker,
-  OnlineHandler
-} from './pools/pool'
+  OnlineHandler,
+  PoolOptions
+} from './pools/abstract-pool'
+export { DynamicClusterPool } from './pools/cluster/dynamic'
+export { FixedClusterPool } from './pools/cluster/fixed'
+export type { ClusterPoolOptions } from './pools/cluster/fixed'
+export type { IPool } from './pools/pool'
 export { DynamicThreadPool } from './pools/thread/dynamic'
 export { FixedThreadPool } from './pools/thread/fixed'
 export type { ThreadWorkerWithMessageChannel } from './pools/thread/fixed'
